editor redesign wip
