+++ apps/desktop/src/routes/editor/ui.tsx
@@ -7,13 +7,14 @@ import { Switch as KSwitch } from "@kobalte/core/switch";
 import { Tooltip as KTooltip } from "@kobalte/core/tooltip";
 import { cva, cx, type VariantProps } from "cva";
 import {
+  mergeProps,
+  splitProps,
   type ComponentProps,
   type JSX,
   type ParentProps,
   type ValidComponent,
-  mergeProps,
-  splitProps,
 } from "solid-js";
+import Tooltip from "~/components/Tooltip";
 import { useEditorContext } from "./context";
 import { TextInput } from "./TextInput";
 
@@ -45,7 +46,7 @@ export function Subfield(
     >
       <span>
         {props.name}
-        {props.required && <span class="text-blue-500 ml-px">*</span>}
+        {props.required && <span class="ml-px text-blue-500">*</span>}
       </span>
       {props.children}
     </div>
@@ -84,7 +85,7 @@ export function Slider(props: ComponentProps<typeof KSlider>) {
       }}
     >
       <KSlider.Track class="h-[0.5rem] relative mx-1">
-        <KSlider.Fill class="absolute bg-blue-100 ui-disabled:bg-gray-300 h-full rounded-full -ml-2" />
+        <KSlider.Fill class="absolute -ml-2 h-full bg-blue-100 rounded-full ui-disabled:bg-gray-300" />
         <KSlider.Thumb
           class={cx(
             "size-[1.25rem] bg-blue-300 -top-1.5 rounded-full outline-none outline-2 outline-offset-2 focus-visible:outline-blue-300 ui-disabled:bg-gray-400"
@@ -120,7 +121,7 @@ export const Dialog = {
           {!props.hideOverlay && (
             <KDialog.Overlay class="fixed inset-0 z-50 bg-[#000]/80 ui-expanded:animate-in ui-expanded:fade-in ui-closed:animate-out ui-closed:fade-out" />
           )}
-          <div class="fixed inset-0 z-50 flex items-center justify-center">
+          <div class="flex fixed inset-0 z-50 justify-center items-center">
             <KDialog.Content
               class={cx(
                 "z-50 divide-y text-sm rounded-[1.25rem] overflow-hidden border border-gray-200 bg-gray-50 min-w-[22rem] ui-expanded:animate-in ui-expanded:fade-in ui-expanded:zoom-in-95 origin-top ui-closed:animate-out ui-closed:fade-out ui-closed:zoom-out-95",
@@ -204,7 +205,7 @@ export function MenuItem<T extends ValidComponent = "button">(
       class={cx(
         props.class,
         "flex flex-row shrink-0 items-center gap-[0.375rem] px-[0.675rem] py-[0.375rem] rounded-[0.5rem] outline-none text-nowrap overflow-hidden text-ellipsis w-full max-w-full",
-        "text-[0.875rem] text-gray-400 disabled:text-gray-400 ui-highlighted:bg-gray-100 ui-highlighted:text-gray-500"
+        "text-[0.875rem] text-gray-400 disabled:text-gray-400 ui-highlighted:bg-gray-200 ui-highlighted:text-gray-500"
       )}
     />
   );
@@ -243,16 +244,16 @@ export function MenuItemList<T extends ValidComponent = "div">(
 const editorButtonStyles = cva(
   [
     "group flex flex-row items-center px-[0.375rem] gap-[0.375rem] h-[2rem] rounded-[0.5rem] text-[0.875rem]",
-    "focus-visible:outline outline-2 outline-offset-2 transition-colors duration-100",
-    "disabled:bg-gray-100 disabled:text-gray-400",
+    "focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 transition-colors duration-100",
+    "disabled:opacity-50 disabled:text-gray-400",
   ],
   {
     variants: {
       variant: {
         primary:
-          "text-gray-500 enabled:hover:ui-not-pressed:bg-gray-100 ui-expanded:bg-gray-100 outline-blue-300",
+          "text-gray-500 enabled:hover:ui-not-pressed:bg-gray-200 ui-expanded:bg-gray-200 outline-blue-300 focus:bg-gray-200",
         danger:
-          "text-gray-500 enabled:hover:ui-not-pressed:bg-gray-100 ui-expanded:bg-red-300 ui-pressed:bg-red-300 ui-expanded:text-gray-50 ui-pressed:text-gray-50 outline-red-300",
+          "text-gray-500 enabled:hover:ui-not-pressed:bg-gray-200 ui-expanded:bg-red-300 ui-pressed:bg-red-300 ui-expanded:text-gray-50 ui-pressed:text-gray-50 outline-red-300",
       },
     },
     defaultVariants: { variant: "primary" },
@@ -263,9 +264,9 @@ const editorButtonLeftIconStyles = cva("transition-colors duration-100", {
   variants: {
     variant: {
       primary:
-        "text-gray-400 enabled:group-hover:not-ui-group-disabled:text-gray-500 ui-group-expanded:text-gray-500",
+        "text-gray-500 enabled:group-hover:not-ui-group-disabled:text-gray-500 ui-group-expanded:text-gray-500",
       danger:
-        "text-gray-400 enabled:group-hover:text-gray-500 ui-group-expanded:text-gray-50 ui-group-pressed:text-gray-50",
+        "text-gray-500 enabled:group-hover:text-gray-500 ui-group-expanded:text-gray-50 ui-group-pressed:text-gray-50",
     },
   },
   defaultVariants: { variant: "primary" },
@@ -273,8 +274,12 @@ const editorButtonLeftIconStyles = cva("transition-colors duration-100", {
 
 type EditorButtonProps<T extends ValidComponent = "button"> =
   PolymorphicProps<T> & {
+    children?: JSX.Element | string;
     leftIcon?: JSX.Element;
     rightIcon?: JSX.Element;
+    tooltipText?: string;
+    comingSoon?: boolean;
+    rightIconEnd?: boolean;
   } & VariantProps<typeof editorButtonStyles>;
 
 export function EditorButton<T extends ValidComponent = "button">(
@@ -282,22 +287,60 @@ export function EditorButton<T extends ValidComponent = "button">(
 ) {
   const [local, cvaProps, others] = splitProps(
     mergeProps({ variant: "primary" }, props) as unknown as EditorButtonProps,
-    ["children", "leftIcon", "rightIcon"],
+    [
+      "children",
+      "leftIcon",
+      "rightIcon",
+      "tooltipText",
+      "comingSoon",
+      "rightIconEnd",
+    ],
     ["class", "variant"]
   );
 
-  return (
-    <Polymorphic
-      as="button"
-      {...others}
-      class={editorButtonStyles({ ...cvaProps, class: cvaProps.class })}
-    >
+  const buttonContent = (
+    <>
       <span class={editorButtonLeftIconStyles({ variant: cvaProps.variant })}>
         {local.leftIcon}
       </span>
-      <span>{local.children}</span>
-      <span class="text-gray-400">{local.rightIcon}</span>
-    </Polymorphic>
+      {local.children && <span>{local.children}</span>}
+      {local.rightIcon && (
+        <span class={local.rightIconEnd ? "ml-auto" : ""}>
+          {local.rightIcon}
+        </span>
+      )}
+    </>
+  );
+
+  return (
+    <>
+      {local.tooltipText || local.comingSoon ? (
+        <Tooltip content={local.comingSoon ? "Coming soon" : local.tooltipText}>
+          <Polymorphic
+            as="button"
+            {...others}
+            class={cx(
+              editorButtonStyles({ ...cvaProps, class: cvaProps.class }),
+              local.rightIconEnd && "justify-between"
+            )}
+            disabled={local.comingSoon}
+          >
+            {buttonContent}
+          </Polymorphic>
+        </Tooltip>
+      ) : (
+        <Polymorphic
+          as="button"
+          {...others}
+          class={cx(
+            editorButtonStyles({ ...cvaProps, class: cvaProps.class }),
+            local.rightIconEnd && "justify-between"
+          )}
+        >
+          {buttonContent}
+        </Polymorphic>
+      )}
+    </>
   );
 }
