+++ apps/desktop/src/routes/editor/utils.ts
@@ -11,3 +11,18 @@ export function formatTime(secs: number, fps?: number) {
 
   return str;
 }
+
+import { getCurrentWindow, ProgressBarStatus } from "@tauri-apps/api/window";
+import { createEffect } from "solid-js";
+
+export function createProgressBar(progress: () => number | undefined) {
+    const currentWindow = getCurrentWindow();
+  
+    createEffect(() => {
+      const p = progress();
+      console.log({ p });
+      if (p === undefined)
+        currentWindow.setProgressBar({ status: ProgressBarStatus.None });
+      else currentWindow.setProgressBar({ progress: Math.round(p) });
+    });
+  }
