+++ apps/desktop/src/utils/titlebar-state.ts
@@ -1,8 +1,8 @@
-import { createStore } from "solid-js/store";
-import type { JSX } from "solid-js";
+import type { UnlistenFn } from "@tauri-apps/api/event";
 import { getCurrentWindow } from "@tauri-apps/api/window";
 import { type as ostype } from "@tauri-apps/plugin-os";
-import type { UnlistenFn } from "@tauri-apps/api/event";
+import type { JSX } from "solid-js";
+import { createStore } from "solid-js/store";
 
 export interface TitlebarState {
   height: string;
@@ -14,6 +14,7 @@ export interface TitlebarState {
   minimizable: boolean;
   closable: boolean;
   border: boolean;
+  backgroundColor: string | null;
   transparent: boolean;
 }
 
@@ -27,6 +28,7 @@ const [state, setState] = createStore<TitlebarState>({
   minimizable: true,
   closable: true,
   border: true,
+  backgroundColor: null,
   transparent: false,
 });
 
@@ -54,5 +56,5 @@ async function initializeTitlebar(): Promise<UnlistenFn | undefined> {
   });
 }
 
-export { setState as setTitlebar, initializeTitlebar };
+export { initializeTitlebar, setState as setTitlebar };
 export default state;
