+++ packages/ui-solid/src/auto-imports.d.ts
@@ -8,8 +8,10 @@ export {}
 declare global {
   const IconCapArrows: typeof import('~icons/cap/arrows.jsx')['default']
   const IconCapAudioOn: typeof import('~icons/cap/audio-on.jsx')['default']
+  const IconCapAuto: typeof import('~icons/cap/auto.jsx')['default']
   const IconCapBlur: typeof import('~icons/cap/blur.jsx')['default']
   const IconCapCamera: typeof import('~icons/cap/camera.jsx')['default']
+  const IconCapCaptions: typeof import('~icons/cap/captions.jsx')['default']
   const IconCapChevronDown: typeof import('~icons/cap/chevron-down.jsx')['default']
   const IconCapCircleCheck: typeof import('~icons/cap/circle-check.jsx')['default']
   const IconCapCirclePlus: typeof import('~icons/cap/circle-plus.jsx')['default']
@@ -25,12 +27,14 @@ declare global {
   const IconCapFilmCut: typeof import('~icons/cap/film-cut.jsx')['default']
   const IconCapFrameFirst: typeof import('~icons/cap/frame-first.jsx')['default']
   const IconCapFrameLast: typeof import('~icons/cap/frame-last.jsx')['default']
+  const IconCapGauge: typeof import('~icons/cap/gauge.jsx')['default']
   const IconCapHotkeys: typeof import('~icons/cap/hotkeys.jsx')['default']
   const IconCapImage: typeof import('~icons/cap/image.jsx')['default']
   const IconCapInfo: typeof import('~icons/cap/info.jsx')['default']
   const IconCapInset: typeof import('~icons/cap/inset.jsx')['default']
   const IconCapInstant: typeof import('~icons/cap/instant.jsx')['default']
   const IconCapLayout: typeof import('~icons/cap/layout.jsx')['default']
+  const IconCapLink: typeof import('~icons/cap/link.jsx')['default']
   const IconCapLogo: typeof import('~icons/cap/logo.jsx')['default']
   const IconCapLogoFull: typeof import('~icons/cap/logo-full.jsx')['default']
   const IconCapLogoFullDark: typeof import('~icons/cap/logo-full-dark.jsx')['default']
